Refactor components/mapper to utils/mapper
