+++ src/utils/mapper/adaptive-code-syntax.ts
@@ -50,4 +50,3 @@ export const getConditionSyntax = (condition:string) : string => {
     code+="};";
     return code;
 };
-
